+++ subsys/net/lib/lwm2m/CMakeLists.txt
@@ -6,6 +6,7 @@ zephyr_include_directories(.)
 
 zephyr_library_sources(
     lwm2m_engine.c
+    lwm2m_observation.c
     lwm2m_obj_security.c
     lwm2m_obj_server.c
     lwm2m_obj_device.c
